UI Construction
Storyboard now includes a Navigation Controller, as well as a Tab View Controller instead of linking buttons to present screens modally. This change was made in order to help lessen the load on the navigation stack.

Implemented "UserDefaults" in order to store a user's data locally. This will aid me in achieving a more personalized UI, as well as making API requests cleaner. Using UserDefaults, I am now able to pass information into the Tab View easily.

My next goal is to enable the app to recognize when the user is logged in vs. when they're not in order to change the initial state of the app.